% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree10000}
\alias{tree10000}
\title{Example Tree data}
\format{
A data frame with 10,000 rows and 6 columns
\describe{
\item{Tree}{unique tree ID number}
\item{BS}{binary presence of bark splits}
\item{DBH}{diameter at breast height, cm}
\item{DB}{percent crown dieback, 5-100\%}
\item{V}{vigor rating with 1 the healthiest and 5 half of crown dead}
\item{WP}{binary presence of woodpecker activity}
}
}
\source{
randomly generated by FSMM authors
}
\usage{
tree10000
}
\description{
An example data set of 10,000 trees randomly generated to provide example function for FSMM.
}
\keyword{datasets}
